+++ Jabberpoint_Sourcecode_Students/src/jabberPoint/loaders/LoaderFactory.java
@@ -1,0 +1,12 @@
+package jabberPoint.loaders;
+
+public abstract class LoaderFactory {
+
+    public static DemoPresentation createDemoPresentation(){
+        return new DemoPresentation();
+    }
+
+    public static XMLAccessor createXMLAccessor(){
+        return new XMLAccessor();
+    }
+}