+++ Jabberpoint_Sourcecode_Students/src/jabberPoint/navigation/AboutBox.java
@@ -1,4 +1,4 @@
-package jabberPoint;
+package jabberPoint.navigation;
 
 import java.awt.Frame;
 import javax.swing.JOptionPane;
@@ -13,7 +13,7 @@ import javax.swing.JOptionPane;
  * @version 1.6 2014/05/16 Sylvia Stuurman
  */
 
-public class AboutBox {
+public abstract class AboutBox {
 	public static void show(Frame parent) {
 		JOptionPane.showMessageDialog(parent,
 				"JabberPoint is a primitive slide-show program in Java(tm). It\n" +