+++ Jabberpoint_Sourcecode_Students/src/jabberPoint/navigation/MenuFactory.java
@@ -1,0 +1,14 @@
+package jabberPoint.navigation;
+
+import jabberPoint.presentation.SlideViewerComponent;
+import jabberPoint.presentation.SlideViewerFrame;
+
+public abstract class MenuFactory {
+    public static KeyController createKeyController(SlideViewerComponent slideViewerComponent){
+        return new KeyController(slideViewerComponent);
+    }
+
+    public static MenuController createMenuController(SlideViewerFrame frame){
+        return new MenuController(frame);
+    }
+}