+++ Jabberpoint_Sourcecode_Students/src/jabberPoint/presentation/Presentation.java
@@ -1,4 +1,4 @@
-package jabberPoint;
+package jabberPoint.presentation;
 
 import java.util.ArrayList;