+++ changes.md
@@ -22,6 +22,13 @@
 - DataType XMLAccessor naar SaveAble en LoadAble voor expandability
 - opening a new presentation updates the parent frame
 - presentation.setSlideNumber(0); verplaatst naar XMLACCESSor loadFile omdat slidenumber altijd 0 moet zijn na het laden van een nieuwe presentation
+- Constructor opgedeeld
 
 ## SlideViewerComponent
 - update() frame update naar Frame verplaatst
+
+## DemoPresentation
+- loadfile opgedeeld
+
+## aboutbox
+- aboutbox abstract